refactor(mcp): consolidate the response layer — 12 files to 9, out of the size baseline
internal/mcp is the shared MCP protocol layer: it owns the canonical
Succeed/Fail/ParseArgs/LenientUnmarshal vocabulary that five tool packages
call. Splitting it into subpackages would change the import path in every
consumer, so this shrinks it by consolidation instead — the import path is
untouched and no caller changes.

Three files go:

- response_markdown.go was a placeholder with no code. Its MarkdownTable and
  Truncate were removed in e3347917 (Truncate deduped into internal/util,
  MarkdownTable dead) and nothing replaced them. The 0.8.0 pre-release review
  already flagged it as dead (docs/0.8.0-pre-release-bugs.md, finding 7).
- response.go held only a "Layout:" comment after f2eac984 extracted its
  contents. The layout note now lives in doc.go, where package-level
  orientation belongs, and covers all eight files rather than four.
- response_builders.go (Succeed/SucceedText/Fail/ParseArgs) folds into the
  file that already holds TextResponse/JSONResponse/SafeMarshal. Those two
  sets are one concern — "turn a summary plus payload into a wire response" —
  and the history agrees: both were touched by the same DRY sweep that created
  the builders.

response_json.go takes over the freed response.go name, since it is no longer
only about JSON marshaling.

No exported symbol added, removed, or renamed. No behavior change: the moved
function bodies are byte-identical.
