--- internal/mcp/response_markdown.go
@@ -1,3 +1,0 @@
-// Purpose: Placeholder — Markdown formatting for MCP text content blocks.
-// Why: Separates Markdown formatting from JSON and content-block construction.
-package mcp
